// frontend/src/domain/types/application.ts
export interface ApplicationFormData {
    fullName: string;
    dateOfBirth: string;
    gender: string;
    education: string;
    otherInfo: string;
    health: string;
    email: string;
    phone: string;
    address: string;
    emergencyContact: string;
    program: string;
    testScores: string;
    recommendations: string;
    essays: string;
    financialInfo: string;
    documents: string;
    declaration: string;
    payment: string;
  }